[Improvement][Worker] Do not verify the status of yarn application in ShellCommandExecutor (#5564)
