+++ dolphinscheduler-server/src/main/java/org/apache/dolphinscheduler/server/worker/task/AbstractYarnTask.java
@@ -16,6 +16,11 @@
  */
 package org.apache.dolphinscheduler.server.worker.task;
 
+import org.apache.dolphinscheduler.common.Constants;
+import org.apache.dolphinscheduler.common.enums.ExecutionStatus;
+import org.apache.dolphinscheduler.common.thread.Stopper;
+import org.apache.dolphinscheduler.common.utils.HadoopUtils;
+import org.apache.dolphinscheduler.common.utils.StringUtils;
 import org.apache.dolphinscheduler.dao.entity.TaskInstance;
 import org.apache.dolphinscheduler.server.entity.TaskExecutionContext;
 import org.apache.dolphinscheduler.server.utils.ProcessUtils;
@@ -23,10 +28,25 @@ import org.apache.dolphinscheduler.service.bean.SpringApplicationContext;
 import org.apache.dolphinscheduler.service.process.ProcessService;
 import org.slf4j.Logger;
 
+import java.io.*;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import static org.apache.dolphinscheduler.common.Constants.EXIT_CODE_FAILURE;
+import static org.apache.dolphinscheduler.common.Constants.EXIT_CODE_SUCCESS;
+
 /**
  *  abstract yarn task
  */
 public abstract class AbstractYarnTask extends AbstractTask {
+  /**
+   * rules for extracting application ID
+   */
+  protected static final Pattern APPLICATION_REGEX = Pattern.compile(Constants.APPLICATION_REGEX);
+
   /**
    *  process task
    */
@@ -55,14 +75,123 @@ public abstract class AbstractYarnTask extends AbstractTask {
     try {
       // SHELL task exit code
       CommandExecuteResult commandExecuteResult = shellCommandExecutor.run(buildCommand());
+      // get appIds from log
+      List<String> appIds = getAppIds(taskExecutionContext.getLogPath());
       setExitStatusCode(commandExecuteResult.getExitStatusCode());
-      setAppIds(commandExecuteResult.getAppIds());
+      // if yarn task , yarn state is final state
+      if (commandExecuteResult.getExitStatusCode() == EXIT_CODE_SUCCESS){
+        setExitStatusCode(isSuccessOfYarnState(appIds) ? EXIT_CODE_SUCCESS : EXIT_CODE_FAILURE);
+      }
+      setAppIds(String.join(Constants.COMMA, appIds));
       setProcessId(commandExecuteResult.getProcessId());
     } catch (Exception e) {
       logger.error("yarn process failure", e);
       exitStatusCode = -1;
       throw e;
     }
+  }
+
+  /**
+   * check yarn state
+   *
+   * @param appIds application id list
+   * @return is success of yarn task state
+   */
+  public boolean isSuccessOfYarnState(List<String> appIds) {
+    boolean result = true;
+    try {
+      for (String appId : appIds) {
+        while(Stopper.isRunning()){
+          ExecutionStatus applicationStatus = HadoopUtils.getInstance().getApplicationStatus(appId);
+          logger.info("appId:{}, final state:{}",appId,applicationStatus.name());
+          if (ExecutionStatus.FAILURE.equals(applicationStatus) ||
+                  ExecutionStatus.KILL.equals(applicationStatus)) {
+            return false;
+          }
+          if (ExecutionStatus.SUCCESS.equals(applicationStatus)){
+            break;
+          }
+          Thread.sleep(Constants.SLEEP_TIME_MILLIS);
+        }
+      }
+    } catch (Exception e) {
+      logger.error(String.format("yarn applications: %s  status failed ", appIds.toString()),e);
+      result = false;
+    }
+    return result;
+
+  }
+
+
+  /**
+   * get app links
+   *
+   * @param logPath log path
+   * @return app id list
+   */
+  private List<String> getAppIds(String logPath) {
+    List<String> logs = convertFile2List(logPath);
+
+    List<String> appIds = new ArrayList<>();
+    /**
+     * analysis log?get submited yarn application id
+     */
+    for (String log : logs) {
+      String appId = findAppId(log);
+      if (StringUtils.isNotEmpty(appId) && !appIds.contains(appId)) {
+        logger.info("find app id: {}", appId);
+        appIds.add(appId);
+      }
+    }
+    return appIds;
+  }
+
+  /**
+   * convert file to list
+   * @param filename file name
+   * @return line list
+   */
+  private List<String> convertFile2List(String filename) {
+    List lineList = new ArrayList<String>(100);
+    File file=new File(filename);
+
+    if (!file.exists()){
+      return lineList;
+    }
+
+    BufferedReader br = null;
+    try {
+      br = new BufferedReader(new InputStreamReader(new FileInputStream(filename), StandardCharsets.UTF_8));
+      String line = null;
+      while ((line = br.readLine()) != null) {
+        lineList.add(line);
+      }
+    } catch (Exception e) {
+      logger.error(String.format("read file: %s failed : ",filename),e);
+    } finally {
+      if(br != null){
+        try {
+          br.close();
+        } catch (IOException e) {
+          logger.error(e.getMessage(),e);
+        }
+      }
+
+    }
+    return lineList;
+  }
+
+  /**
+   * find app id
+   * @param line line
+   * @return appid
+   */
+  private String findAppId(String line) {
+    Matcher matcher = APPLICATION_REGEX.matcher(line);
+    if (matcher.find()) {
+      return matcher.group();
+    }
+    return null;
   }
 
   /**
